refactor: validation id

+++ controller/post.controller.js
@@ -1,52 +1,59 @@
 const Post = require("../models/post.model");
 
+//GET
 const getAllPosts = async (req, res) => {
     const allPosts = await Post.find({});
     res.json(allPosts);
 };
-
+//POST
 const addPost = async (req, res) => {
     const body = req.body;
-    // console.log("esto es lo que nos llega",body)
-    // const {title, content} = body;
     try {
-        // const existPost = await Post.findOne({title, content});
-        // if(existPost){
-        //     return res.status(400).json({
-        //         ok: false,
-        //         msj: "post exist",
-        //     })
-        // }
-        const postDevTo = new Post(req.body);
+        const postDevTo = new Post(body);
         await postDevTo.save();
-        res.json({
+        res.status(200).json({
             ok: true,
             postDevTo,
         });
     } catch (err) {
-        console.log("este es un errer ", err);
+        // console.log("este es un error ", err);
         res.status(500).json({
             ok: false,
-            msj: "error in post save",
+            msj: "bad post post",
+            error: err,
         });
     }
 };
+//GETBYID
 const getById = async (req, res) => {
     const id = req.params.id;
     try {
         const postById = await Post.findById(id);
-
-        res.json(postById);
-    } catch (error) {
-        console.error(error);
-        res.end();
+        //Validacion de Id-----------
+        if (!postById) {
+            //Validacion de Post en dado caso que un post no sea encontrado mediante la ID
+            return res.status(404).json({
+                //Respuesta con estatis 404
+                ok: false,
+                msj: "el post no existe", //Mensaje de la respuesta
+            });
+        }
+        res.status(200).json(postById);
+    } catch (err) {
+        // console.log("este es un error ", err);
+        res.status(500).json({
+            ok: false,
+            msj: "bad gerByID.",
+            error: err,
+        });
     }
 };
-
+//PUT
 const editPost = async (req, res) => {
     const id = req.params.id;
     try {
         const postID = await Post.findById(id); //Busqueda de post con el ID para verificar que ID sea valido
+        //Validacion de Id-----------
         if (!postID) {
             //Validacion de Post en dado caso que un post no sea encontrado mediante la ID
             return res.status(404).json({
@@ -66,17 +73,29 @@ const editPost = async (req, res) => {
             ok: true,
             updatePost,
         });
-    } catch (error) {
-        console.log(error);
+    } catch (err) {
+        // console.log(err);
         res.status(500).json({
             ok: false,
             msj: "bad update.",
+            error: err,
         });
     }
 };
+//DELETE
 const deletePost = async (req, res) => {
+    const id = req.params.id;
     try {
-        const id = req.params.id;
+        const postID = await Post.findById(id); //Busqueda de post con el ID para verificar que ID sea valido
+        //Validacion de Id-----------
+        if (!postID) {
+            //Validacion de Post en dado caso que un post no sea encontrado mediante la ID
+            return res.status(404).json({
+                //Respuesta con estatis 404
+                ok: false,
+                msj: "el post no existe", //Mensaje de la respuesta
+            });
+        }
         const deletedPost = await Post.findByIdAndDelete(id);
         res.statusCode = 200;
         res.json({
